lscpu-cputype: fix missing fields for a new cputype while parsing
The cputype parsing reads lines from the /proc/cpuinfo in a streaming
fashion, and a new cputype is created only if we find there is a different
CPU attribute, the newly created cputype does not contain attributes that are
the same as the previous cputype.

Copy back these same CPU attributes when creating a new cputype struct.

Signed-off-by: Sung-Chi Li <alan81920@gmail.com>