setpriv: consistently use "<caps>" to indicate a list of capabilities
The existing "<caps,...>" did not indicate that each capability name
must be preceded by "+" or "-".  Also, to indicate that the list needs
to be separated by commas, one would have to use "<cap>[,...]", not
"<cap,...>".  So... instead of trying to cram all that info in there,
just say "<caps>", like for --bounding-set.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>