mkswap: fix includes
It is preferable to use #ifdefs in code that relies on available
features, rather than solely depending on the presence of header
files. This is especially important if the functionality of FIEMAP
relies on multiple include files. Additionally, note that we also need
linux/fs.h for S_NOCOW_FL, regardless of FIEMAP.

Signed-off-by: Karel Zak <kzak@redhat.com>