usage: mention also the missing KiB and MiB as permissible suffixes
Also, slightly reword the phrase to be more grammatical,
and make it into a full sentence that ends in a period.

(By the way, commit f1970cc557 is somewhat problematic: it requires
that the relevant occurrences of "<num>" and "<size>" in the option
descriptions are now always translated in the same manner, but the
translator may be unaware of this.  For Dutch, the latter was the
case for `ipcmk` and `zramctl`. :/ )

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>