zramctl: wait for device being initialized and unlocked by udevd
systemd-udevd takes a lock during processing the uevent for a block
device. The kernel refuses 'reset' attribute for zram device is written
when the device node is opened. Hence, during systemd-udevd processes a
uevent for zram device, we cannot write 'reset' attribute.
Let's wait for the device being initialized and unlocked by udevd.

Note, there still exists a race window, as we need to release the lock
before writing 'reset' attribute. But, the situation should be better
now.