scriptreplay: mark literal values in the man page in bold, not italic
Also, mention which keys can be used to speed up, slow down, and pause
the replay.  Shorten the description of '-s' as it is deprecated.  And
improve the description of '-t': it does not override the _timingfile_
positional argument, but takes the place of it, makes it unneeded.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>
(cherry picked from commit df2353c92ed5ea632826a4b81f22cb38f944407c)