zramctl: restore normal description indentations in the usage text
Commit 19c12e63b1 from four months ago changed the indentation of *all*
the option descriptions in the usage text.  That was the wrong thing to
do, as it invalidates all existing translations for no good reason.

When one of the options is (or needs to become) overlong, then the thing
to do is: to put the description of this option on the next line.

So, restore the indentation of the descriptions to what they were
before commit 19c12e63b1, and wrap the description of -p.

Also, make the description of -b fit within 80 columns, and sort the
options better: alphabetically, but single letter before long option.

CC: LiviaMedeiros <livia@cirno.name>
Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>