setarch: (man) correct the markup of the synopsis and of two options
Drop the second line of the synopsis as it is unneeded (it is covered
by the first line since argument "arch" became optional) and distracts
from the similarity/contrast between the other two lines.

Mark "arch" in the new second synopsis line in italics, as it is a
placeholder, not a literal.

Also, improve some wording, reduce redundancy by reshuffling the --pid
option, and remove a redudant -v from an example.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>