Do not use strerror on shared libraries
Using strerror is not safe in shared libraries as it is unknown
what the calling apps may do with the storage behind the scenes.
Using %m is ok, except in contexts that need to be as-safe.

(cherry picked from commit c5850c36e14085cddf8db9e19d447d928fa153d0)