lib/loopdev: open loop control device read-only if possible
Opening the control device read-write to trigger udev events is a hack.
It works around a kernel bug which fails to send uevents directly from the
kernel by triggering a full udev reprobe.
While it works to trigger the events, it is problematic.
The udev reprobe will asynchronously recreate all partition devices which
requires an ungly locking and synchronization scheme after losetup has
completed to guarantee the usability of the partition devices.

The kernel has been fixed to properly emit these uevents in
commit 7ed2a771b5fb ("loop: properly send KOBJ_CHANGED uevent for disk device").

Drop the workaround.

This is a partial revert of
commit ced1142d6f25 ("losetup: cleanup device node modes").

Signed-off-by: Thomas Weißschuh <thomas.weissschuh@linutronix.de>