rtcwake: (man) reduce an inflated table to sane proportions
Asciidoctor puts an unneeded blank line into every table cell (this
is fixed in git [1], but not released yet), and will unnecessarily
wrap text in a cell when the text is slightly longer than average.

Replace the table with a block of preformatted text, and replace
the "+5min" example with a true format that covers most cases,
and give some examples of that format after the block.

[1] https://github.com/asciidoctor/asciidoctor/commit/9cb73f8c9bee

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>