terminal-colors.d: (man) reduce two tables to succinct lists
Before the move to asciidoctor, these tables were succinct lists in the
man page.  Change the asciidoc text to make them into lists again.

Use {nbsp} on one line to avoid weird groff behavior for \? when .ss
is defined to be zero-width (as asciidoctor does), and reshuffle two
paragraphs to avoid an asciidoctor bug that adds a spurious \fP after
a monospace span defined with double backticks.

Also, correct some markup in a few earlier paragraphs.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>