lscpu: (man) don't refer to a missing section, and improve some wordings
Three years ago commit aa049eabb3 removed the COLUMNS section but forgot
to remove the reference to it.  Replace the reference with a referral to
the --help text.

In the bargain, join two paragraphs that belong together, improve some
wordings, and sort two options.

This fixes #2590 (https://github.com/util-linux/util-linux/issues/2590).
Reported-by: Fabien Malfoy

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>