libmount: (monitor) clean up internal names
In future changes, we need to process active monitor entries in entry
type-specific code to return data from events.

This requires a slight change in the function's logic, so it makes
sense to rename the function according to the new logic, meaning
changes --> active, verify --> process.

The patch also replaces bit fields with booleans.

Signed-off-by: Karel Zak <kzak@redhat.com>