bash-completion: prefer "command lsblk" to "lsblk" to avoid aliases
The user may define an alias lsblk to specify the default formatting,
such as -T, which would conflict with lsblk -r used in the Bash
completion settings. In the Bash completion settings, to prevent the
calls of lsblk -r ... from being replaced with lsblk -T -r ... and
failing, lsblk should be called using command lsblk.