mount: add --exclusive command line option
The new option ensures that the kernel does not reuse existing
superblock.

The new option is available to non-root users as it does not affect
the superblock itself or any other aspects of the mount process. It
makes the current mounting more restrictive, so it makes sense to
allow it for non-root users.

Signed-off-by: Karel Zak <kzak@redhat.com>