lsfd: revise the code disabling hyperlinks
In the original code, the code was at file.c.  This change lifts up
the code to upper layer.  As a result the fill_column method of struct
file becomes simple.

As a side effect. this change disables hyperlinks for the entries of
UNKN type (e.g. eventpool, or pidfd). This is a bug fix.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>