swapon: control column header tooltips with --annotation
Whether column header names should be annotated can be specified
with the --annotation option which takes the following three
arguments: 'auto', 'never' and 'always', where the former is the
default and makes it so that annotations are only added when the
output device is a terminal. If the option is omitted the 'auto'
behavior should be expected.

Signed-off-by: Christian Goeschel Ndjomouo <cgoesc2@wgu.edu>