include/c.h: add USAGE_LIST_COLUMNS_OPTION() macro
With this new macro we no longer have to redundantly specify
`fputs(_(" -H, --list-columns...` in the usage() function of ls-like tools, as it
will be consolidated to the USAGE_LIST_COLUMNS_OPTION() macro.

USAGE_LIST_COLUMNS_OPTION() aligns the option description to the other descriptions
in the usage() output in the same way as USAGE_HELP_OPTIONS().

Signed-off-by: Christian Goeschel Ndjomouo <cgoesc2@wgu.edu>