swapon: (man page) use "defaults" (plural)
A lot of documentation on the Internet seems to assume "defaults" is the
/correct/ default value when no other options are intended.
Documentation/example.files/fstab does not have an entry for swap, but
it shows "defaults" for other file systems.

It seems prudent to align on a single variant, at least in the
documentation, even if both are accepted by swapon.

Signed-off-by: Chris Hofstaedtler <zeha@debian.org>